Add angular-materials to the shared UI module
added angular materials to shared UI module to be used by the other ui engineers should they use any materials
