+++ libs/client/company-explore/feature/src/lib/company-explore-feature/company-explore-content/company-explore-content.component.spec.ts
@@ -1,5 +1,5 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { CompanyExploreMaterialModule } from '../../materials/company-explore-material.module';
+import { UiModule } from '@graduates/client/shared/assets/ui';
 import { CompanyExploreCardComponent } from './company-explore-card/company-explore-card.component';
 
 import { CompanyExploreContentComponent } from './company-explore-content.component';
@@ -10,7 +10,7 @@ describe('CompanyExploreContentComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      imports:[CompanyExploreMaterialModule],
+      imports:[UiModule],
       declarations: [ CompanyExploreContentComponent,CompanyExploreCardComponent ]
     })
     .compileComponents();
